Add ability to run processes in container
For now, uses 'runc exec' to run processes in the container. To make
sure we have a container running that we can 'exec' processes in, uses a
dummy process which prints 'Pid 1 Running' on stdout (and waits forever)
as the Pid 1.

Imports iodaemon and process_tracker mostly unmodified from garden-linux,
but to avoid in-memory state (for now) generates ever-increasing process
ids shared by all proceses.