Wire in Networking
Networks are created as network namespaces in /var/run/netns and passed to the
containerizer as a path to the network namespace which the containerizer can
enter. The network namespace is set up before container creation by directly
entering the namespace ourselves in go code (which means we get nice logs etc).

Signed-off-by: Julian Friedman <julz.friedman@uk.ibm.com>