Use pre-start hook for networking
this is needed for user namespaces. Instead of creating a network
namespace in the network component and returning the path, we now have
the network component return a pre-start hook (path and args) that can
set up the networking for a container. This means that the networking is
just a regular OCI pre-start hook and the bundle works as a unit without
guardian.

Removes ForeignNetworkAdaptor since generic hooks should be able to
replace it.

Signed-off-by: Julian Friedman <julz.friedman@uk.ibm.com>
Signed-off-by: Gareth Smith <gsmith@pivotal.io>