Avoid leaking iodaemon sockets when spawned process dies uncleanly
In the case where the iodaemon-spawned process dies with a SIGKILL we never
recieve an EOF on Stdout/Stderr. This causes a slow leak of pipes and sockets
since the FDs are never closed. To avoid this we defensively add a timeout in
case the pipes are not closed even though an exit status was recieved. This
appears to be the only way to make this work without resorting to low-level
polling of FDs.

[#112062877]