Work around inability to build Linux CGO on Mac
The combination of CGO and GOOS=Linux breaks compilation on a Mac, which
has the result of breaking Vim's compilation error checking. We've
worked around this by adding a non-CGO implementation of our single
CGO package, which will be used when CGO_ENABLED=0. We're adding that
env var setting to our Vim alias in the garden-dotfiles repo.

[finishes #139469133]

Signed-off-by: Gareth Clay <gclay@pivotal.io>