User processes survive gdn server restart
Previously, user processes without TTYs were sent a sigpipe if they
tried to write to stdout/stderr when the gdn server was down. This was
because there were no processes on the system with read FDs open for
these named pipes.

Dadoo now keeps a stdin write and stdout/stderr read FDs open, but never
writes or reads them. This allows the user process to read and write to
its end of the stdio pipes.

[#139930811]

Signed-off-by: Georgi Sabev <georgethebeatle@gmail.com>