Relax goroutine leak test constraint
The test currently checks that the number of goroutines before and after
creation and destruction of a container are equal. It seems like when we
collect the pre-create goroutine count sometimes there is an extra
goroutine, possibly from something during server startup. Either way, we
can relax this constraint as the test should only ensure the number of
goroutines doesn't go up.

[#145129027]

Signed-off-by: Feroz Jilla <fjilla@pivotal.io>