Convert preloaded+layer image URLs to OCI images
Only when using image plugins.

An `image` dir is created under the container's directory in the depot,
and OCI image spec-compliant metadata files are written. These files
reference 2 layers: the lower one is a file:// URL referring to the
locally-stored base rootFS, and the supper one is an http(s):// URL
referring to the some tarball (e.g. a Cloud Foundry droplet).

An annotation is used to indicate to the image plugin where in the
rootFS to untar the tarball.

[#148289649]

Signed-off-by: Will Pragnell <wpragnell@pivotal.io>
Signed-off-by: Roberto Jimenez Sanchez <jszroberto@gmail.com>