Rootless /dev/pts owned by gid 0 rather than 5
Runc validates that no mount points in rootless containers have non-zero
uid or gid. When we want to enable TTYs in rootless containers we will
probably have to address this, but for now we just set the gid to 0
instead of 5 (5 being the arbitrary gid we select as the owner of
/dev/pts).
In the container, this mount point appears to be owned by the
unprivileged host gid, rather than 0.

[#150122955]