Increase timeout for ip netns add in a test
We also reverted test suite's global timeout back to the default. The
idea here is to reduce signal to noise and to leave only the ip netns add
with a very large timeout. If we still hit this very large timeout, then
we'll know there's something further that needs investigating.

[#146163395]

Signed-off-by: Ed King <eking@pivotal.io>