Executable to open socket FDs for guardian
socket2me will be used in bosh to open the unix socket as root, before
changing uid and gid and exec-ing the guardian server. In this way
rootless guardian can use a socket that its user doesn't own. This is
valuable because the gdn server user is also container root.

[#151125161]

Signed-off-by: Craig Furman <cfurman@pivotal.io>