Fix quota leak for the 1st rootfs created in the store
As seen in  https://www.pivotaltracker.com/story/show/152030731,
when using the overlay-xfs driver, the first rootfs created would be
assigned the projectID 1.
But calls to stats, or other creates, would endup re-assigning that
projectID to the QuotaControl itself, and reseting it to zero, causing
the quota in that rootfs to not be enforced.
This will stop projectID 1 from being used for rootfs, and leave it only
for the internal QuotaControl.

[fixes #152277099]

Signed-off-by: Georgi Sabev <georgethebeatle@gmail.com>