Flag to disable privileged container creation
Defaults to false when not specified, allowing privileged containers to
be created. In the bosh deployment, we set this flag to false when
experimental_rootless_mode is enabled.

This flag can be enabled by standalone gdn binary users (e.g. Concourse)
to disable creation of privileged containers even on rootful
deployments.

[#151996438]