Use runc run -d on Linux to create container
This change allows us to use the tiny init.c program as the container
init process. We have had to extract "createCmd" and
"createCmdExtraArgs" in order to support runc, winc and peas. Right now,
in order to create a container, runc uses "runc run -d", peas
use "runc run" and winc uses "winc create". This is a temporary
situation, and the plan is to eventually only support "runc/winc run
-d".

[#149413675]

Signed-off-by: Julia Nedialkova <julianedialkova@hotmail.com>
Signed-off-by: Ed King <eking@pivotal.io>