Use dadoo/execrunner for peas
The motivation of this user story is to get container.Run() to return an
error when the executable doesn't exist when the process is a pea, just
as it does for regular processes.

We refactored dadoo/execrunner so that they could be used for all
processes on Linux.

This probably gets us several other pieces of functionality for free,
but nothing is promised until we write tests.

[#151713778]

Signed-off-by: Georgi Sabev <georgethebeatle@gmail.com>
Signed-off-by: Craig Furman <cfurman@pivotal.io>