Write exec log files to a temp dir on Windows
Before this change, a pea process would write its log file to a
subdirectory of the main container's directory in the depot. This causes a
problem when the main container is deleted. The deletion of the
container's depot directory will fail because the exec process for the pea will
have an open handle to its log file. This causes guardian to incorrectly report the
number of containers, as it counts the un-deleted directories in the
deport (even though the actual containers are gone)

Writing the log file to a temp dir allows deletion of the container's
depot directory to succeed

[#154635041]

Signed-off-by: Andrew Keesler <akeesler@pivotal.io>