Add the iptablesTimeout to flakey test
It looks like iptables applies a backoff timeout strategy when
attempting to acquire the xtables lock. Bumping the timeout will allow
one more attempt to acquire the lock before failing the test.

This is what the other iptables-related tests are configured to use at
the moment, so it makes sense to set it here as well.

[#154715915]

Signed-off-by: Tom Godkin <tgodkin@pivotal.io>