When creating cgroups, look up the mount source and skip it
This is, basically, a workaround for https://github.com/moby/moby/issues/34584
When creating cgroup directories, check that our cgroup mountpoint isn't
itself a child cgroup; if it is, adjust the paths accordingly so that
the paths (relative to the top cgroup) is correct.  Otherwise we end up
with cgroups that have duplicate sections in the path.