Bump limit on the "no. of logs generated" test
We're actually not totally sold on the value of this test anymore, given
that declarative healthchecks (in which the no. of container processes,
and by extension no. of logs is fewer than before) are enabled by
default in cf-deployment.

But, we want to follow up on this before removing the test permanently.

Signed-off-by: Ed King <eking@pivotal.io>