Stream exec logs on Windows
Instead of passing `winc` a path to a log file, use a pipe and give
`winc` the write end of the pipe. This allows the logs from `winc exec`
to be forwarded as they are written. Previously, the exec logs would
only be forwarded after the exec'd process exited, making debugging
difficult

Notes:
- This relies on winc-release 1.2.0 and later
- This prevents a race condintion where garden.Destroy() would fail if
the `winc exec` command was writing to its log file

[#154291838]

Signed-off-by: Amin Jamali <ajamali@pivotal.io>
Signed-off-by: Natalie Arellano <narellano@pivotal.io>