Add container.Attach for windows
This is impemented by creating pipes for stdin, stdout, and stderr,
giving one side of each to the winc process and keeping the other side
in the garden server.

For multiplexing output, use the same strategy as on linux.

For input, each client (whether from Run or Attach) gets a copy of the
write side of the stdin handle.

[#157004409]

Signed-off-by: Ed King <eking@pivotal.io>
Signed-off-by: Amin Jamali <ajamali@pivotal.io>
Signed-off-by: Natalie Arellano <narellano@pivotal.io>
Signed-off-by: Sam Smith <sesmith177@gmail.com>