Bump defaultEventuallyTimeout in dadoo tests
These sometimes flake on overloaded CI workers, so bumping the limit up
to a minute. Ideally I'd like there to not be a default timeout and to
have slowmobius notify us of any hanging tests, but there isn't a
particularly elegant way to do that in Ginkgo, so I'm bumping the limit
for now. We can revist if we still see timeout-related flakes here.