Remove unused test file.
It seems ginkgo fails to compile test suites that only have _linux parts
on Windows. We previously added an empty _test.go to get around this,
but it seems this is leaving crap in C:\Windows\Temp\go-link-...\ on
each run. This eventually causes the concourse worker to become too
fragmented.

Instead, we'll ignore the nerd suite in the script that runs these tests
on Windows and I'll raise an issue(/PR?) to ginkgo to report the issue.

[#156607857]