Initialise grootfs stores only when needed in GQTS
This change prevents a loop device leak. Prior to this change the
grootfs stores were initalised in the global BeforeEach. Cleanup of the
stores occurs in each test suite's local AfterEach. This meant that we
were accidentally skipping over cleanup on tests we were Skiping (e.g.
the containerd tests when container is not enabled). This left a bunch
of loop devices around which caused other tests further down the line to
fail.

[#158969527]

Co-authored-by: Claudia Beresford <cberesford@pivotal.io>