fix int overflow on 32bit platforms
Linux uses u32 as the {u,g}id map structures[1] but golang incorrectly
uses ints[2]. Since changing the standard library is hard this commit
only changes the parsing code to use the correct types and truncates the
result to MaxInt.

The effect is that on 32bit platforms the namespaces will have at most 2
billion available uid instead of 4 billion.

[1] https://github.com/torvalds/linux/blob/v4.17/include/linux/user_namespace.h#L17-L21
[2] https://github.com/golang/go/blob/5c11480631f5654e9e6937ff08c453660138c64d/src/syscall/exec_linux.go#L14-L20

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>