List any open file in the store when its deletion fails
This uses `lsof` to recursively list any open files contained in the
store we've just failed to delete. `lsof <path-to-dir>` is not enough to
do this though: it will just list "processes that have it open as a file
or as a process-specific directory". Do recursively list open files in a
directory, we need to use `lsof +D <path-to-dir>`. I've also added `-x`
to follow symlinks. Hopefully this doesn't turn out to be too slow!

http://man7.org/linux/man-pages/man8/lsof.8.html