Extract the network depot
The network depot holds files that are needed by the networker to setup
the container network. Previously these files were stored in the
container depot. Now we want to get rid of the container depot when
running with containerd, therefore we split the depot into container
depot and network depot.

[#155622234]

Co-authored-by: Danail Branekov <danailster@gmail.com>