Do not log empty output from image plugins
We have seen in our environments that any image plugin command that
attaches the ReLogger will still print out a log line if there is empty
output. This is especially misleading in logs b/c it does not parse the
empty output as a lager log line and then prints out an error level log
(with no error, but the log message has '.error' appended). This change
should also reduce the total log volume of guardian (e.g. in a log file
from a Windows cell with 82982 total lines, there were 42459 that
matched 'image-plugin.*error' and had an empty output string in the log
data)

Co-authored-by: Amin Jamali <ajamali@pivotal.io>