Remove as much resources as possible on destroy
Previously gardener would stop on the first error during the destroy
routine thus leaving debris behind. With this commit we make best
effort to delete the resources that are safe to delete. For example, we
delete the bundle last if and only if the rest of the container
artifacts are deleted.

[#159850215]

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>
Co-authored-by: Claudia Beresford <cberesford@pivotal.io>