Run clean on create asynchronously
`grootfs create --with-clean` will now spawn `grootfs clean` as a
subprocess so that create does not have to wait for the cleanup to
complete. The subprocess will write its log to a separate file
configured with the --with-clean-log flag. If this flag is not set the
subprocess will log to stderr.

[#166282986]

Co-authored-by: Giuseppe Capizzi <gcapizzi@pivotal.io>
Co-authored-by: Julia Nedialkova <julianedialkova@hotmail.com>
Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>