Fix pea metrics on Windows
The metrics endpoint was broken for peas when using OCI runtimes that
don't report the container creation date in the output of the `state`
command.

When the output of `state` does not include the container creation date,
we inspect the depot for the container pidfile and use its modified time
instead. The depot inspection fails for peas though, thus the error
above.

I have fixed this by searching for the pea pidfile in the `processes`
folder of all existing containers, which should succeed in the case of
peas.

[#167796160]

Co-authored-by: Giuseppe Capizzi <gcapizzi@pivotal.io>