Wait 2 minutes for the roots mount to disappear
There is no point in umounting the rootfs in the fail handler because we
know that unmounting the rootfs within grootfs succeeded (there are no
unmount errors in the logs)
Instead, we just wait for 2 minutes for the rootfs to disappear from the
mount table. If we confirm that this is the case indeed, then we will
know that unmounting is "racy" and we could think how to handle that
raciness with grace.

[#168988942]

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>