Invoke containerd client close on gdn stop
The containerd client maintains a GRPC connection to the containerd
server which is nice to be closed when no longer needed so that we avoid
connection leaks. Closing the connection makes sense in scenarios
when gdn restarts and connects to a previously started containerd
server, such as gdn standalone and garden GQT restart tests.

In the CF bosh world closing the connection probably is not that
beneficial as stopping the garden job also stops the containerd server.

[#169368014]