Unthrottle containers with CPU usage < CPU entitlement
When a punished container (previously moved to the bad cgroup) keeps
calm and its overall CPU usage goes below its entitlement, it is
being forgiven and is released back to the good cgroup. Of course, it
will be punished to the bad croup again once it gets naughty and its
overall CPU usage exceeds its CPU entitlement.

[#168137341]