Run process using supplementary groups
Previously an empty slice of SGIDs would be passed in the process spec.
We think the intention was to ensure supplementary groups of the running
garden process were not inherited into the container process. However,
this does not happen, and it is safe to pass the SGIDs calculated from
/etc/group.

In rootless mode, we cannot set supplementary groups, so an empty slice
is passed as before.

[#174096808]

Co-authored-by: Kieron Browne <kbrowne@pivotal.io>
Co-authored-by: Mario Nitchev <marionitchev@gmail.com>