Always nest peas cgroups under the sandbox ones
With runc later than v1.0.0-rc91, deleting a container signals all the
processes in the cgroup. If the pea container and the sandbox container
share the same cgroup, this results into killing the sandbox init
process when deleting the pea. This change ensures that the pea
container cgroup is nested into the sandbox one, thefore deleting the
pea does not kill the sandbox init process.

[#174751988]

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>