Switch to the io.containerd.runc.v2 runtime.
Some regressions that we have noticed:

1. Init tasks don't have associated info anymore. We had to rewrite
tests that relied on it.
2. Getting a non-existing process doesn't result in a containerd error
anymore. A process with Status = "unknown" is returned instead. We have
added a workaround so that Nerd still returns a ProcessNotFoundError,
and filed an issue with containerd:
https://github.com/containerd/containerd/issues/4632
We hope to be able to remove our workaround if the issue gets fixed.

[#174426706]

Co-authored-by: Kieron Browne <kbrowne@pivotal.io>