Ensure containerd IO is completely copied on `Wait`
When the the containerd process exit channel gets closed, this only
indicates that the process is done, but does not mean that the
background routines that are copying stdout/stderr are done.

Users of Garden API expect that when a process is waited for, its output
is immediately avaialble, that is why we need to also ensure that
containerd async copying is done.

Co-authored-by: Danail Branekov <danailster@gmail.com>