Containers() returns all containers when garden.state is all
By default if no garden.state property is provided Containers() is
returning containers in `created` state.
If the value is set to "all" it will remove the property and not filter
by it. This is done for backwards compatibility with old API. Ideally,
we would always return all containers in the Containers() call, but
since this was added some time ago we have to be backwards compatible.

Signed-off-by: Maria Shaldybin <mariash@vmware.com>