Set container properties before creating container
and everything it depends on, like volumizer, networker. So that we can
fetch this container even if it was partially created for cleanup. E.g.
rep is setting container owner and only cleans up containers that it
created as an owner. If container is partially created (e.g. setting up
network fails). We want to be able to remove it later, even if immediate
destroy fails (e.g. external networker is down)

Signed-off-by: Renee Chu <reneec@vmware.com>