Update quota tests for ginkgo v2
In ginkgo v2, BeforeSuite now runs once per worker, instead
of once ever. As a result we need to shard the pool of XFS MountPoints
for each worker, rather than use one pool for all of them.

Also we need to make sure to return resources back to the pool when
done, since we didn't always do this previously.