Pin Microsoft/hcsshim to 0.10.0-rc.8
An earlier commit updated github.com/containerd/containerd to 1.7.2.
That version of containerd rquires hcsshim 0.10.0-rc.8.
However, between 0.10.0-rc.8 and the officially release 0.10.0,
hcsshim changed some of its interfaces/function signatures,
so updating to 0.10.0 broke containerd.

Until containerd accommodates the interface change from hcsshim,
or reverts to an older version of hcsshim, we'll need to keep this
pin in place.

Signed-off-by: Marc Paquette <mpaquette@vmware.com>
Signed-off-by: David Sabeti <sabetid@vmware.com>