Use /proc dir in test to force xattr fetch error
This test depends on failing to fetch xattrs (extended attributes) from
the directory where a container's `image_info` file will be saved.
Specifically, the xattr being used here is the XFS project ID. The test
validates the behavior when the `tardis` binary cannot fetch the project
ID.

This is the second time we've had to change the directory used in the
test. The original implementation of the test used `/tmp`, but in the
Bionic update, `/tmp` started having xattrs set on it, causing the test
to fail. We switched to `/dev` next, but the most recent update to the
Jammy stemcell includes a major Linux Kernel bump which also allos
xattrs to fetched from `/dev`.

Switching to `/proc` might be kicking the can down the road. We should
plan to explore how we can create a temp directory for this test, such
that fetching the xattrs on that directory fails as needed for this
test.