Remove pin on otel to 1.15.1
When attempting to update the other go mod dependencies, we are
receiving the following error:
```
go: code.cloudfoundry.org/guardian/tools imports
	github.com/containerd/containerd/cmd/containerd imports
	github.com/containerd/containerd/cmd/containerd/builtins imports
	github.com/containerd/containerd/tracing/plugin imports
	go.opentelemetry.io/otel/sdk/resource imports
	go.opentelemetry.io/otel/semconv/v1.21.0: module go.opentelemetry.io/otel@latest found (v1.21.0, replaced by go.opentelemetry.io/otel@v1.15.1), but does not contain package go.opentelemetry.io/otel/semconv/v1.21.0
```

Upon removing the pin to the 1.15 version and then updating the go mod
dependencies, we see the vendor is finally working without issue. We'd
like to attempt to remove this pinned version entirely.

Signed-off-by: Maria Shaldybin <mariash@vmware.com>