Improve external networker error message
The error message captures both the error and the action that led to the error.
Because the most common action that causes this error is the 'down' action (meaning, network teardown),
a common error message was 'external networker down,' leading many people to beleive that the entire vxlan-policy-agent
was no longer running.

The new error message tries to make clear that 'down' is the action, to avoid this misunderstanding

Signed-off-by: Josh Russett <josh.russett@broadcom.com>